move cmdhelper module to otaclient_common
