+++ src/otaclient/app/boot_control/_jetson_cboot.py
@@ -25,7 +25,7 @@ import subprocess
 from pathlib import Path
 from typing import Generator, Optional
 
-from otaclient import cmdhelper
+from otaclient_common import cmdhelper
 from otaclient.app import errors as ota_errors
 from otaclient.app.boot_control._common import OTAStatusFilesControl, SlotMountHelper
 from otaclient.app.boot_control._jetson_common import (
