+++ src/otaclient/app/boot_control/_jetson_common.py
@@ -29,7 +29,7 @@ from typing import Any, NamedTuple, Optional
 from pydantic import BaseModel, BeforeValidator, PlainSerializer
 from typing_extensions import Annotated, Literal, Self
 
-from otaclient import cmdhelper
+from otaclient_common import cmdhelper
 from otaclient_common.common import copytree_identical, write_str_to_file_sync
 
 logger = logging.getLogger(__name__)
